manage_canvas_by_state#27
