+++ package.json
@@ -20,6 +20,8 @@
     "react-scripts": "5.0.0",
     "react-scroll-horizontal": "^1.6.6",
     "superagent": "^7.0.1",
+    "sweetalert2": "^11.4.4",
+    "sweetalert2-react-content": "^4.2.0",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
