+++ src/style.css
@@ -28,7 +28,7 @@ img {
   text-align: center;
 }
 .heading {
-  margin: 5px;
+  margin: 10px 10px 0 5px;
 }
 .margin-left {
   margin-left: 10px
@@ -676,6 +676,7 @@ video {
   border-style: solid;
   border-color: #000;
   margin-right: 0;
+  margin-top: 0.5rem;
 }
 .tag{
   border-left: solid 6px #2d8fdd;/*左側の線*/
